+++ packages/metalsmith-plugins/packages/metalsmith-link-checker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.2 / 2021-11-24
+
+- Fix the `ignore` option broken in v0.1.0.
+- Re-implement a HTTP/HTTPS network cache removed in v0.1.0.
+- De-duplicate links before checking.
+
 ## v0.1.0 / 2021-11-24
 
 - Print the error next to failed links.